Refactor validationn DSL to more idiomatic Typeclass pattern
