+++ validation/src/main/scala/hmda/validation/dsl/PredicateDefaults.scala
@@ -1,31 +1,6 @@
 package hmda.validation.dsl
 
-trait CommonDsl {
-  implicit class Subject[T](data: T) {
-    private def test(predicate: Predicate[T]): Result = {
-      predicate.validate(data) match {
-        case true => Success()
-        case false => Failure(predicate.failure)
-      }
-    }
-
-    private def testNot(predicate: Predicate[T]): Result = {
-      predicate.validate(data) match {
-        case true => Failure(predicate.failure)
-        case false => Success()
-      }
-    }
-
-    def is(predicate: Predicate[T]): Result = {
-      test(predicate)
-    }
-
-    def not(predicate: Predicate[T]): Result = {
-      testNot(predicate)
-    }
-
-  }
-
+object PredicateDefaults {
   def equalTo[T](that: T): Predicate[T] = new Predicate[T] {
     override def validate: (T) => Boolean = _ == that
     override def failure: String = s"not equal to $that"
