+++ validation/src/main/scala/hmda/validation/rules/lar/syntactical/S010.scala
@@ -1,10 +1,13 @@
 package hmda.validation.rules.lar.syntactical
 
 import hmda.model.fi.lar.LoanApplicationRegister
-import hmda.validation.dsl.{ CommonDsl, Result }
+import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
 
 object S010 extends EditCheck[LoanApplicationRegister] {
+
+  import hmda.validation.dsl.PredicateSyntax._
+  import hmda.validation.dsl.PredicateDefaults._
 
   //Hardcoded for now
   val larId = 2
