+++ validation/src/main/scala/hmda/validation/rules/lar/syntactical/S011.scala
@@ -1,10 +1,14 @@
 package hmda.validation.rules.lar.syntactical
 
 import hmda.model.fi.lar.LoanApplicationRegister
-import hmda.validation.dsl.{ CommonDsl, Result }
+import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
 
 object S011 extends EditCheck[Iterable[LoanApplicationRegister]] {
+
+  import hmda.validation.dsl.PredicateDefaults._
+  import hmda.validation.dsl.PredicateSyntax._
+
   def apply(lars: Iterable[LoanApplicationRegister]): Result = {
     lars.size is greaterThan(0)
   }
