+++ validation/src/main/scala/hmda/validation/rules/lar/syntactical/S020.scala
@@ -1,10 +1,14 @@
 package hmda.validation.rules.lar.syntactical
 
 import hmda.model.fi.lar.LoanApplicationRegister
-import hmda.validation.dsl.{ CommonDsl, Result }
+import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
 
 object S020 extends EditCheck[LoanApplicationRegister] {
+
+  import hmda.validation.dsl.PredicateDefaults._
+  import hmda.validation.dsl.PredicateSyntax._
+
   //Hardcoded for now
   val agencyCodes: List[Int] = List(1, 2, 3, 5, 7, 9)
