+++ validation/src/main/scala/hmda/validation/rules/lar/validity/V255.scala
@@ -5,6 +5,10 @@ import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
 
 object V255 extends EditCheck[LoanApplicationRegister] {
+
+  import hmda.validation.dsl.PredicateDefaults._
+  import hmda.validation.dsl.PredicateSyntax._
+
   def apply(lar: LoanApplicationRegister): Result = {
     lar.actionTakenType is containedIn(1 to 8)
   }
