+++ validation/src/main/scala/hmda/validation/rules/lar/validity/V310.scala
@@ -6,6 +6,9 @@ import hmda.validation.rules.EditCheck
 
 object V310 extends EditCheck[LoanApplicationRegister] {
 
+  import hmda.validation.dsl.PredicateDefaults._
+  import hmda.validation.dsl.PredicateSyntax._
+
   def apply(lar: LoanApplicationRegister): Result = {
     lar.applicant.race1 is containedIn(1 to 7)
   }
