+++ validation/src/main/scala/hmda/validation/rules/lar/validity/V340.scala
@@ -6,6 +6,9 @@ import hmda.validation.rules.EditCheck
 
 object V340 extends EditCheck[LoanApplicationRegister] {
 
+  import hmda.validation.dsl.PredicateDefaults._
+  import hmda.validation.dsl.PredicateSyntax._
+
   def apply(lar: LoanApplicationRegister): Result = {
     lar.purchaserType is containedIn(0 to 9)
   }
