+++ validation/src/main/scala/hmda/validation/rules/lar/validity/V347.scala
@@ -5,6 +5,10 @@ import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
 
 object V347 extends EditCheck[LoanApplicationRegister] {
+
+  import hmda.validation.dsl.PredicateDefaults._
+  import hmda.validation.dsl.PredicateSyntax._
+
   override def name: String = "V347"
 
   def apply(lar: LoanApplicationRegister): Result = {
