+++ validation/src/main/scala/hmda/validation/rules/ts/syntactical/S013.scala
@@ -1,11 +1,14 @@
 package hmda.validation.rules.ts.syntactical
 
 import hmda.model.fi.ts.TransmittalSheet
-import hmda.validation.dsl.{ CommonDsl, Result }
-
+import hmda.validation.dsl.Result
 import scala.concurrent.{ ExecutionContext, Future }
 
-object S013 extends CommonDsl {
+object S013 {
+
+  import hmda.validation.dsl.PredicateDefaults._
+  import hmda.validation.dsl.PredicateSyntax._
+
   def apply(ts: TransmittalSheet, timestamp: Long): Result = {
     val t = ts.timestamp
     t is greaterThan(timestamp)
@@ -15,4 +18,7 @@ object S013 extends CommonDsl {
     val t = ts.timestamp
     fTimestamp.map(time => t is greaterThan(time))
   }
+
+  def name: String = "S013"
+
 }
