+++ validation/src/main/scala/hmda/validation/rules/ts/validity/V105.scala
@@ -8,6 +8,10 @@ import hmda.validation.rules.EditCheck
  Respondent name, address, city, state, and zip code must not = blank
  */
 object V105 extends EditCheck[TransmittalSheet] {
+
+  import hmda.validation.dsl.PredicateDefaults._
+  import hmda.validation.dsl.PredicateSyntax._
+
   def apply(ts: TransmittalSheet): Result = {
     val respondent = ts.respondent
     val respName = respondent.name
