+++ validation/src/main/scala/hmda/validation/rules/ts/validity/V155.scala
@@ -1,10 +1,14 @@
 package hmda.validation.rules.ts.validity
 
 import hmda.model.fi.ts.TransmittalSheet
-import hmda.validation.dsl.{ RegexDsl, Result }
+import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
 
-object V155 extends EditCheck[TransmittalSheet] with RegexDsl {
+object V155 extends EditCheck[TransmittalSheet] {
+
+  import hmda.validation.dsl.PredicateSyntax._
+  import hmda.validation.dsl.PredicateRegEx._
+
   def apply(ts: TransmittalSheet): Result = {
     val contact = ts.contact
     val email = contact.email
